requirements.txt
githubpy==1.1.0
pickleDB==0.6.2
simplejson==3.13.2
pyodbc
flask
flask-restful
pyodbc
flask-cors
